Alterações em Exceptions, validações

+++ src/main/java/br/com/jsf_pfproject/djp/service/AgendamentoService.java
@@ -7,7 +7,8 @@ import javax.inject.Inject;
 
 import br.com.jsf_pfproject.djp.dao.AgendamentoDAO;
 import br.com.jsf_pfproject.djp.dao.DAO;
-import br.com.jsf_pfproject.djp.exception.AgendamentoException;
+import br.com.jsf_pfproject.djp.exception.RemoverException;
+import br.com.jsf_pfproject.djp.exception.ValidationException;
 import br.com.jsf_pfproject.djp.model.Agendamento;
 import br.com.jsf_pfproject.djp.utility.ValidationUtil;
 
@@ -30,7 +31,7 @@ public class AgendamentoService implements Serializable {
 	public void salvar(Agendamento agendamento) {
 
 		if (!ValidationUtil.dataAgendamentoValida(agendamento.getDataAgendamento())) {
-			throw new AgendamentoException("Não é possível agendar para uma data e hora anterior a atual");
+			throw new ValidationException("Não é possível agendar para uma data e hora anterior a atual");
 		}
 		
 		// Valida se o agendamento passado por parâmetro é nulo, se for é um novo agendamento
@@ -42,11 +43,10 @@ public class AgendamentoService implements Serializable {
 			// Valido se data ou professor desse update foram alterados, se sim, valido se esta data ou professor já possuem
 			// agendamentos vinculados a eles
 			if (!agendamentoTemp.getDataAgendamento().equals(agendamento.getDataAgendamento())
-					|| !agendamentoTemp.getProfessor().getId().equals(agendamento.getProfessor().getId())) {
-				
+					|| !agendamentoTemp.getProfessor().getId().equals(agendamento.getProfessor().getId())) {				
 				
 				validaDataDuplicada(agendamento);
-			}
+			} 
 			
 		} else {
 			validaDataDuplicada(agendamento);
@@ -57,14 +57,20 @@ public class AgendamentoService implements Serializable {
 	}
 
 	public void remover(Agendamento agendamento) {
+		
+		if (agendamento.getId() == null) {
+			throw new RemoverException("Não existe agendamento com esses parâmetros");
+			
+		}		
 		agendamentoDao.remover(Agendamento.class, agendamento.getId());
+		
 	}
 
 	public List<Agendamento> listarTodos() {
 		return agendamentoDao.buscarTodos(Agendamento.class);
 	}
 
-	// 
+
 	/**
 	 * Validação da existência de um agendamento com os mesmos parâmetros data/hora,
 	 * professor.
@@ -72,9 +78,8 @@ public class AgendamentoService implements Serializable {
 	 * @param agendamento
 	 */
 	private void validaDataDuplicada(Agendamento agendamento) {
-		if (agendamentoDAO.existeObjeto(agendamento.getDataAgendamento(), agendamento.getProfessor().getId())) {
-			
-			throw new AgendamentoException("O professor já possui agendamento para esta data e hora");
+		if (agendamentoDAO.existeObjeto(agendamento.getDataAgendamento(), agendamento.getProfessor().getId())) {			
+			throw new ValidationException("O professor já possui agendamento para esta data e hora");
 		}
 	}
 }
